[DROOLS-3321] Add GWT feature to kie-wb-common-command-api
