+++ kie-wb-common-command-api/src/main/java/org/kie/workbench/common/command/client/CommandListener.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2018 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.kie.workbench.common.command.client;
+
+public interface CommandListener<T, V> {
+
+    void onAllow(final T context,
+                 final Command<T, V> command,
+                 final CommandResult<V> result);
+
+    void onExecute(final T context,
+                   final Command<T, V> command,
+                   final CommandResult<V> result);
+
+    void onUndo(final T context,
+                final Command<T, V> command,
+                final CommandResult<V> result);
+}
