+++ kie-wb-common-command-api/src/main/java/org/kie/workbench/common/command/client/registry/DynamicRegistry.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2018 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.kie.workbench.common.command.client.registry;
+
+/**
+ * Mutable registry that can add or remove items at runtime.
+ * @param <T> The type of the item.
+ */
+public interface DynamicRegistry<T> extends Registry<T> {
+
+    void register(final T item);
+
+    boolean remove(final T item);
+}
