+++ kie-wb-common-command-api/src/main/resources/org/kie/workbench/common/command/CommonCommandAPI.gwt.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.7.0//EN"
+    "http://www.gwtproject.org/doctype/2.7.0/gwt-module.dtd">
+
+<module>
+
+  <source path="client"/>
+</module>
